requirements.txt
# Core Dependencies
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.6
python-socketio>=5.11.0
requests>=2.31.0
pandas>=2.2.0  # Updated for better Python 3.11+ support
numpy>=1.26.2

# Data Science
scipy>=1.11.4
scikit-learn>=1.3.2

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
psutil>=5.9.6
cryptography>=42.0.4  # Security: NULL pointer dereference and Bleichenbacher timing oracle patches
schedule>=1.2.1
python-dateutil>=2.8.2
beautifulsoup4>=4.12.3
apscheduler>=3.10.4
markdown>=3.5.1
pytz>=2024.1
PyJWT>=2.8.0

# API Integration
# NOTE: NO SDK - Using raw requests for full control over error handling
# upstox-python-sdk removed - we build custom wrapper per requirements
protobuf>=4.25.0
websocket-client>=1.6.1

# Development & Testing
pytest>=7.4.3
pytest-mock>=3.12.0  # For mocking in unit tests
pytest-cov>=4.1.0    # For code coverage
pytest-asyncio>=0.23.5 # For async tests
black>=23.12.1

# AI Assistant
google-generativeai>=0.8.0
python-telegram-bot>=22.0
transformers>=4.37.0
torch>=2.0.0

# Production Server
gunicorn>=22.0.0  # Security: HTTP request/response smuggling patches
setproctitle>=1.3.3

# Retry Logic
tenacity>=8.2.3

# AI Services
openai>=2.0.0


# NiceGUI Dashboard
nicegui>=1.4.0

# Redis Caching
redis>=5.0.0
flask-caching>=2.1.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23

# Monitoring
prometheus-client>=0.19.0

# Rate Limiting
Flask-Limiter>=3.5.0

# Error Tracking
sentry-sdk[flask]>=1.39.2

# Security & Validation
Flask-WTF>=1.2.1
marshmallow>=3.20.1

# Compression
Flask-Compress>=1.14

# Async Support
aiohttp>=3.13.3  # Security fix: CVE patches for zip bomb, DOS, and directory traversal
